+++ Changelog.md
@@ -19,6 +19,7 @@
 * Add parsing of `PageSetup#paper_size_name`
 * Add parsing `TableProperties#fill`
 * Add parsing `TablePart#TableColumns`
+* Add parsing `Theme#fontScheme`
 
 ### Refactor
 * Store `sdt` as `DocxParagraph#character_styles_array` element